requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==3.2.7
django-allauth==0.51.0
django-cors-headers==3.13.0
django-discover-runner==1.0
django-filter==22.1
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-on-heroku==1.1.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-role-permissions-pr95==2.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
JSON-minify==0.3.0
MarkupSafe==2.1.1
npm==0.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
optional-django==0.1.0
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==1.5
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
